Deduplicate helpers and split oversized source files
Behavior-preserving refactor of the menubar app and cctop-hook sources.

- HookHandler reuses SessionData's sysctl process helpers instead of its
  own copies; SessionData.processExists replaces HookHandler.isPIDAlive.
- Inline session-file decodes use SessionData.fromFile, and
  SessionManager.deriveLifecycle(for:now:) replaces repeated
  SessionLifecyclePolicy.lifecycle calls.
- runFocusCommand replaces five identical Process launchers in the focus
  code; withSessionLock and withSessionLockIfAvailable share one core.
- PanelCoordinator.handle is split into per-mode handlers with shared
  open/close helpers.
- Shared WorktreeCleanupCandidate.State accent colors,
  RecentProject.compactProjectPath, PluginRemoveButton, and
  SettingsSecondaryRow reuse replace duplicated view and model code.
- AppDelegate, HookHandler, SessionData, and SessionManager+Archive are
  split into extension files along their existing MARK sections, which
  removes the file-length, type-length, and complexity lint suppressions
  they needed. New files are registered in both targets where the
  original file was.
