+++ menubar/CctopMenubar/Models/SessionData.swift
@@ -1,4 +1,3 @@
-// swiftlint:disable file_length
 import Foundation
 
 // MARK: - Shared date formatting
@@ -56,135 +55,6 @@ extension JSONDecoder {
 enum HostAppBundleID {
     static let claudeDesktop = "com.anthropic.claudefordesktop"
     static let codexDesktop = "com.openai.codex"
-}
-
-struct TerminalInfo: Codable, Equatable {
-    let program: String
-    let sessionId: String?
-    let tty: String?
-    let bundleId: String?
-    let socket: String? // Remote-control socket (e.g. KITTY_LISTEN_ON)
-    let multiplexer: MultiplexerInfo?
-    let binaryPaths: [String: String]?
-
-    enum CodingKeys: String, CodingKey {
-        case program
-        case sessionId = "session_id"
-        case tty
-        case bundleId = "bundle_id"
-        case socket
-        case multiplexer
-        case binaryPaths = "binary_paths"
-    }
-
-    init(program: String = "", sessionId: String? = nil, tty: String? = nil,
-         bundleId: String? = nil, socket: String? = nil, multiplexer: MultiplexerInfo? = nil,
-         binaryPaths: [String: String]? = nil) {
-        self.program = program
-        self.sessionId = sessionId
-        self.tty = tty
-        self.bundleId = bundleId
-        self.socket = socket
-        self.multiplexer = multiplexer
-        self.binaryPaths = binaryPaths
-    }
-}
-
-/// Identifies a terminal multiplexer (cmux, herdr, zellij, tmux) hosting the session.
-/// Each variant carries exactly the fields needed for its focus command.
-enum MultiplexerInfo: Codable, Equatable {
-    /// cmux focus-surface --workspace $workspaceId --surface $surfaceId
-    case cmux(socket: String, workspaceId: String, surfaceId: String?, paneId: String?, binaryPath: String?)
-    /// herdr agent focus $paneId (with HERDR_SOCKET_PATH=$socket)
-    case herdr(socket: String, paneId: String, binaryPath: String?)
-    /// zellij --session $sessionName action focus-pane-id $paneId
-    case zellij(sessionName: String, paneId: String, binaryPath: String?)
-    /// tmux -S $socket select-window -t $paneId && tmux -S $socket select-pane -t $paneId
-    case tmux(socket: String, paneId: String, binaryPath: String?)
-
-    private enum CodingKeys: String, CodingKey {
-        case name
-        case sessionName = "session_name"
-        case workspaceId = "workspace_id"
-        case surfaceId = "surface_id"
-        case paneId = "pane_id"
-        case socket
-        case binaryPath = "binary_path"
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let name = try container.decode(String.self, forKey: .name)
-        let binaryPath = try container.decodeIfPresent(String.self, forKey: .binaryPath)
-        switch name {
-        case "cmux":
-            self = .cmux(
-                socket: try container.decode(String.self, forKey: .socket),
-                workspaceId: try container.decode(String.self, forKey: .workspaceId),
-                surfaceId: try container.decodeIfPresent(String.self, forKey: .surfaceId),
-                paneId: try container.decodeIfPresent(String.self, forKey: .paneId),
-                binaryPath: binaryPath
-            )
-        case "herdr":
-            self = .herdr(
-                socket: try container.decode(String.self, forKey: .socket),
-                paneId: try container.decode(String.self, forKey: .paneId),
-                binaryPath: binaryPath
-            )
-        case "zellij":
-            self = .zellij(
-                sessionName: try container.decode(String.self, forKey: .sessionName),
-                paneId: try container.decode(String.self, forKey: .paneId),
-                binaryPath: binaryPath
-            )
-        case "tmux":
-            self = .tmux(
-                socket: try container.decode(String.self, forKey: .socket),
-                paneId: try container.decode(String.self, forKey: .paneId),
-                binaryPath: binaryPath
-            )
-        default:
-            throw DecodingError.dataCorruptedError(
-                forKey: .name, in: container,
-                debugDescription: "Unknown multiplexer: \(name)"
-            )
-        }
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        switch self {
-        case .cmux(let socket, let workspaceId, let surfaceId, let paneId, let binaryPath):
-            try container.encode("cmux", forKey: .name)
-            try container.encode(socket, forKey: .socket)
-            try container.encode(workspaceId, forKey: .workspaceId)
-            try container.encodeIfPresent(surfaceId, forKey: .surfaceId)
-            try container.encodeIfPresent(paneId, forKey: .paneId)
-            try container.encodeIfPresent(binaryPath, forKey: .binaryPath)
-        case .herdr(let socket, let paneId, let binaryPath):
-            try container.encode("herdr", forKey: .name)
-            try container.encode(socket, forKey: .socket)
-            try container.encode(paneId, forKey: .paneId)
-            try container.encodeIfPresent(binaryPath, forKey: .binaryPath)
-        case .zellij(let sessionName, let paneId, let binaryPath):
-            try container.encode("zellij", forKey: .name)
-            try container.encode(sessionName, forKey: .sessionName)
-            try container.encode(paneId, forKey: .paneId)
-            try container.encodeIfPresent(binaryPath, forKey: .binaryPath)
-        case .tmux(let socket, let paneId, let binaryPath):
-            try container.encode("tmux", forKey: .name)
-            try container.encode(socket, forKey: .socket)
-            try container.encode(paneId, forKey: .paneId)
-            try container.encodeIfPresent(binaryPath, forKey: .binaryPath)
-        }
-    }
-}
-
-extension MultiplexerInfo {
-    var isCmux: Bool {
-        if case .cmux = self { return true }
-        return false
-    }
 }
 
 struct SubagentInfo: Codable, Equatable {
@@ -584,85 +454,5 @@ extension SessionData {
         case "agent": return "Spawning: \(detail.prefix(30))"
         default: return "\(tool): \(detail.prefix(30))"
         }
-    }
-}
-
-// MARK: - Process Liveness
-
-extension SessionData {
-    static func processInfo(pid: UInt32) -> kinfo_proc? {
-        var info = kinfo_proc()
-        var size = MemoryLayout<kinfo_proc>.size
-        var mib: [Int32] = [CTL_KERN, KERN_PROC, KERN_PROC_PID, Int32(pid)]
-        let result = sysctl(&mib, UInt32(mib.count), &info, &size, nil, 0)
-        guard result == 0, size > 0 else { return nil }
-        return info
-    }
-
-    static func processStartTime(pid: UInt32) -> TimeInterval? {
-        guard let info = processInfo(pid: pid) else { return nil }
-        return startTime(from: info)
-    }
-
-    var isAlive: Bool {
-        guard let pid else { return false }
-        guard kill(Int32(pid), 0) == 0 || errno == EPERM else { return false }
-        guard let info = Self.processInfo(pid: pid) else { return false }
-
-        if let stored = pidStartTime {
-            let current = Self.startTime(from: info)
-            if abs(stored - current) > 1.0 { return false }
-        }
-
-        // A live process that is identifiably a DIFFERENT harness's binary cannot be this
-        // session's host: the capture-time parent walk can adopt a foreign harness PID, and
-        // rapid PID reuse can land within the 1s start-time tolerance (issue #155).
-        if Self.isForeignHarnessComm(Self.commandName(from: info), source: source) { return false }
-
-        // Suspended (Ctrl+Z) or orphaned (PPID=1) processes are unreachable
-        if info.kp_proc.p_stat == 4 { return false }
-        if info.kp_eproc.e_ppid == 1 { return false }
-
-        return true
-    }
-
-    private static func startTime(from info: kinfo_proc) -> TimeInterval {
-        let tv = info.kp_proc.p_starttime
-        return TimeInterval(tv.tv_sec) + TimeInterval(tv.tv_usec) / 1_000_000
-    }
-
-    /// Kernel-reported executable basename (`p_comm`, truncated to MAXCOMLEN).
-    /// p_comm stores MAXCOMLEN chars + NUL; the +1 keeps a full-length comm's
-    /// terminator inside the rebound region.
-    static func commandName(from info: kinfo_proc) -> String {
-        var proc = info.kp_proc
-        return withUnsafePointer(to: &proc.p_comm) { ptr in
-            ptr.withMemoryRebound(to: CChar.self, capacity: Int(MAXCOMLEN) + 1) { cStr in
-                String(cString: cStr)
-            }
-        }
-    }
-
-    static func processCommandName(pid: UInt32) -> String? {
-        processInfo(pid: pid).map { commandName(from: $0) }
-    }
-
-    /// Maps a process name to the harness that owns that binary. Codex also ships
-    /// arch-suffixed binaries (`codex-aarch64-apple-darwin`, truncated by MAXCOMLEN),
-    /// hence the prefix match.
-    static func harnessOwningComm(_ comm: String) -> String? {
-        if comm == "claude" { return ccSource }
-        if comm == "codex" || comm.hasPrefix("codex-") { return codexSource }
-        if comm == "opencode" { return opencodeSource }
-        if comm == "pi" { return piSource }
-        return nil
-    }
-
-    /// True when the process name belongs to a DIFFERENT harness than this session's.
-    /// Conservative: an unrecognized name proves nothing (claude could run under a
-    /// wrapper). Legacy nil sources are Claude Code sessions.
-    static func isForeignHarnessComm(_ comm: String, source: String?) -> Bool {
-        guard let owner = harnessOwningComm(comm) else { return false }
-        return owner != (source ?? ccSource)
     }
 }
