+++ menubar/CctopMenubar/Services/SessionManager+Archive.swift
@@ -1,214 +1,4 @@
-// swiftlint:disable file_length
 import Foundation
-
-/// The external inputs SessionManager consults while deriving session state: the sessions
-/// directory, the Codex/Claude Desktop archive stores, desktop-app liveness, process liveness,
-/// notification and manual-visibility preferences, and the clock. Production uses `.live()`; tests override
-/// individual fields to run the full pipeline against temp directories, stub lookups, and a
-/// deterministic clock. One state-deriving input remains outside this seam:
-/// `adjustPermissionStatus` probes the live process tree (`proc_listchildpids`) directly.
-struct SessionDataSources {
-    var sessionsDir: URL
-    var codexThreads: any CodexThreadStateProviding
-    var claudeDesktopSessions: any ClaudeDesktopSessionStateProviding
-    var desktopAppConnection: DesktopAppConnectionLookup
-    var processAlive: (SessionData) -> Bool
-    var notificationsEnabled: () -> Bool
-    var notificationClient: SessionNotificationClient = .live
-    var manualSessionVisibility: ManualSessionVisibilityStore = .live
-    var now: () -> Date
-
-    /// A function rather than a stored constant so `Config.sessionsDir()` is resolved
-    /// when the caller constructs its sources. The live metadata stores resolve their
-    /// own paths as needed, with short internal caches for repeated reads in one pass.
-    static func live() -> SessionDataSources {
-        SessionDataSources(
-            sessionsDir: URL(fileURLWithPath: Config.sessionsDir()),
-            codexThreads: CodexThreadArchiveLookup(),
-            claudeDesktopSessions: ClaudeDesktopSessionArchiveLookup(),
-            desktopAppConnection: .live,
-            processAlive: { $0.isAlive },
-            notificationsEnabled: { UserDefaults.standard.bool(forKey: "notificationsEnabled") },
-            now: Date.init
-        )
-    }
-}
-
-struct SessionClassificationEvidence {
-    let archivedCodexThreadIDs: Set<String>
-    let missingCodexThreadIDs: Set<String>
-    let codexInternalHelperThreadIDs: Set<String>
-    let uncertainCodexDelegationThreadIDs: Set<String>
-    let contradictoryCodexDelegationThreadIDs: Set<String>
-    let codexExecHelperThreadIDs: Set<String>
-    let archivedClaudeSessionIDs: Set<String>
-}
-
-struct FrozenCodexThreadState: CodexThreadStateProviding {
-    let index: CodexThreadStateIndex?
-
-    func stateIndex(matching threadIDs: Set<String>) -> CodexThreadStateIndex? {
-        guard let index else { return nil }
-        return index.filtered(to: threadIDs)
-    }
-
-    func existingThreadIDs(matching threadIDs: Set<String>) -> Set<String>? {
-        guard let index, index.unknownThreadIDs.isDisjoint(with: threadIDs) else {
-            return nil
-        }
-        return index.existingThreadIDs.intersection(threadIDs)
-    }
-
-    func archivedThreadIDs(matching threadIDs: Set<String>) -> Set<String>? {
-        guard let index else { return [] }
-        guard index.unknownThreadIDs.isDisjoint(with: threadIDs) else {
-            return nil
-        }
-        return index.archivedThreadIDs.intersection(threadIDs)
-    }
-
-    func internalHelperThreadIDs(matching threadIDs: Set<String>) -> Set<String>? {
-        index.map { $0.internalHelperThreadIDs.intersection(threadIDs) } ?? []
-    }
-
-    func execHelperThreadIDs(matching threadIDs: Set<String>) -> Set<String>? {
-        index.map { $0.execHelperThreadIDs.intersection(threadIDs) } ?? []
-    }
-
-}
-
-enum SessionHiddenReason: Equatable {
-    case persistedHidden
-    case autoHidden
-    case archivedCodexThread
-    case missingCodexThread
-    case codexInternalHelper
-    case codexExecHelper
-    case archivedClaudeDesktop
-    case orphanedEndedClaudeDesktop
-    case claudeDesktopStartupPlaceholder
-
-    /// Archived Codex records are explicit cleanup input, never path protection. Helpers and
-    /// durable local hides still represent live ownership and remain protective.
-    var protectsCleanupPath: Bool {
-        switch self {
-        case .persistedHidden, .autoHidden, .codexInternalHelper, .codexExecHelper:
-            return true
-        case .archivedCodexThread, .missingCodexThread, .archivedClaudeDesktop,
-             .orphanedEndedClaudeDesktop, .claudeDesktopStartupPlaceholder:
-            return false
-        }
-    }
-}
-
-enum SessionDisposition: Equatable {
-    case display
-    case hidden(SessionHiddenReason)
-}
-
-struct ClassifiedSessionRecord {
-    let url: URL
-    let candidate: SessionRecord
-    let disposition: SessionDisposition
-}
-
-struct SessionClassificationSnapshot {
-    let records: [ClassifiedSessionRecord]
-    let evidence: SessionClassificationEvidence
-
-    var displayCandidates: [SessionRecord] {
-        records.compactMap { record in
-            guard record.disposition == .display else { return nil }
-            return record.candidate
-        }
-    }
-
-    var autoHiddenSessions: [(URL, SessionData)] {
-        records.compactMap { record in
-            guard case .hidden(.autoHidden) = record.disposition else { return nil }
-            return (record.url, record.candidate.data)
-        }
-    }
-
-    var codexInternalHelperCandidates: [SessionRecord] {
-        records.compactMap { record in
-            guard case .hidden(.codexInternalHelper) = record.disposition else { return nil }
-            return record.candidate
-        }
-    }
-
-    var protectedProjectPathsForCleanup: Set<String> {
-        let displayProtected = SessionIdentityPolicy
-            .dedupedCandidatesByStableKey(displayCandidates)
-            .filter { $0.lifecycleRank != SessionLifecycle.finished.rawValue }
-            .map(\.data.projectPath)
-        let hiddenProtected = records.compactMap { record -> String? in
-            guard case .hidden(let reason) = record.disposition,
-                  reason.protectsCleanupPath,
-                  record.candidate.lifecycleRank != SessionLifecycle.finished.rawValue else {
-                return nil
-            }
-            return record.candidate.data.projectPath
-        }
-        return Set(displayProtected).union(hiddenProtected)
-    }
-
-    var finishedNonDesktopCandidates: [SessionRecord] {
-        displayCandidates.filter {
-            $0.lifecycleRank == SessionLifecycle.finished.rawValue
-                && !$0.data.isCodex
-                && $0.data.hostClass != .desktop
-        }
-    }
-
-    func manualHiddenProjectPaths(_ hiddenSessionIDs: Set<String>) -> Set<String> {
-        Set(records.compactMap { record in
-            guard let cctopSessionID = record.candidate.data.cctopSessionId,
-                  hiddenSessionIDs.contains(cctopSessionID) else { return nil }
-            return record.candidate.data.projectPath
-        })
-    }
-
-    /// Archived conversations can seed worktree cleanup while preserving the session file.
-    /// Finished non-retained sessions continue to contribute through history.
-    var cleanupSources: [SessionDataCleanupSource] {
-        records.compactMap { record in
-            guard case .hidden(let reason) = record.disposition,
-                  reason.emitsCleanupSource,
-                  record.candidate.data.hasCleanupSourcePath else {
-                return nil
-            }
-            return SessionDataCleanupSource(data: record.candidate.data)
-        }
-    }
-
-    var archivedCodexThreadIDs: Set<String> { evidence.archivedCodexThreadIDs }
-    var missingCodexThreadIDs: Set<String> { evidence.missingCodexThreadIDs }
-    var codexInternalHelperThreadIDs: Set<String> { evidence.codexInternalHelperThreadIDs }
-    var uncertainCodexDelegationThreadIDs: Set<String> { evidence.uncertainCodexDelegationThreadIDs }
-    var contradictoryCodexDelegationThreadIDs: Set<String> { evidence.contradictoryCodexDelegationThreadIDs }
-    var codexExecHelperThreadIDs: Set<String> { evidence.codexExecHelperThreadIDs }
-    var archivedClaudeSessionIDs: Set<String> { evidence.archivedClaudeSessionIDs }
-}
-
-private extension SessionHiddenReason {
-    var emitsCleanupSource: Bool {
-        switch self {
-        case .archivedCodexThread, .archivedClaudeDesktop:
-            return true
-        case .persistedHidden, .autoHidden, .missingCodexThread, .codexInternalHelper,
-             .codexExecHelper, .orphanedEndedClaudeDesktop, .claudeDesktopStartupPlaceholder:
-            return false
-        }
-    }
-}
-
-private extension SessionData {
-    var hasCleanupSourcePath: Bool {
-        let path = WorktreeCleanupScanner.standardizedPath(projectPath)
-        return !path.isEmpty && path != "/"
-    }
-}
 
 extension SessionManager {
     func retainedFinishedCleanupSources(
@@ -277,8 +67,7 @@ extension SessionManager {
         guard Self.isLegacyUUIDFilename(url.deletingPathExtension().lastPathComponent) else { return false }
         if !dataSources.manualSessionVisibility.hasStoredHideEvidence {
             try? FileManager.default.removeItem(at: url) // Pre-PID legacy file; no live writer to race.
-        } else if let fileData = try? Data(contentsOf: url),
-                  let data = try? JSONDecoder.sessionDecoder.decode(SessionData.self, from: fileData),
+        } else if let data = try? SessionData.fromFile(path: url.path),
                   !shouldRetainFinishedManualHideEvidence(data, matching: manualHides) {
             try? FileManager.default.removeItem(at: url)
         }
@@ -354,22 +143,13 @@ extension SessionManager {
         var finished: [SessionData] = []
         for url in jsonFiles {
             guard !Self.isLegacyUUIDFilename(url.deletingPathExtension().lastPathComponent),
-                  let fileData = try? Data(contentsOf: url),
-                  let data = try? JSONDecoder.sessionDecoder.decode(SessionData.self, from: fileData),
+                  let data = try? SessionData.fromFile(path: url.path),
                   !data.hidden,
                   !data.shouldAutoHide,
                   data.isCodex || data.hostClass == .desktop else {
                 continue
             }
-            let life = SessionLifecyclePolicy.lifecycle(
-                for: data,
-                hostClass: data.hostClass,
-                processAlive: dataSources.processAlive(data),
-                now: now,
-                windows: Self.lifecycleWindows,
-                desktopAppRunning: Self.desktopAppRunning(for: data, lookup: dataSources.desktopAppConnection)
-            )
-            if life == .finished {
+            if deriveLifecycle(for: data, now: now) == .finished {
                 finished.append(data)
             }
         }
@@ -382,391 +162,6 @@ extension SessionManager {
         if !claudeFinishedIDs.isEmpty {
             _ = dataSources.claudeDesktopSessions.archivedSessionIDs(matching: claudeFinishedIDs)
         }
-    }
-
-    nonisolated static func sessionClassificationSnapshot(
-        in candidates: [SessionRecord],
-        codexThreads: any CodexThreadStateProviding = CodexThreadArchiveLookup(),
-        claudeDesktopSessions: any ClaudeDesktopSessionStateProviding = ClaudeDesktopSessionArchiveLookup(),
-        now: Date = Date()
-    ) -> SessionClassificationSnapshot {
-        let sessions = candidates.map(\.data)
-        let claudeMetadata = claudeDesktopMetadataSnapshot(in: sessions, claudeDesktopSessions: claudeDesktopSessions)
-        return sessionClassificationSnapshot(
-            in: candidates,
-            sessions: sessions,
-            claudeMetadata: claudeMetadata,
-            codexThreads: codexThreads,
-            now: now
-        )
-    }
-
-    nonisolated static func sessionClassificationSnapshot(
-        in candidates: [SessionRecord],
-        claudeMetadata: ClaudeDesktopSessionMetadataSnapshot?,
-        codexThreads: any CodexThreadStateProviding = CodexThreadArchiveLookup(),
-        now: Date = Date()
-    ) -> SessionClassificationSnapshot {
-        sessionClassificationSnapshot(
-            in: candidates,
-            sessions: candidates.map(\.data),
-            claudeMetadata: claudeMetadata,
-            codexThreads: codexThreads,
-            now: now
-        )
-    }
-
-    private nonisolated static func sessionClassificationSnapshot(
-        in candidates: [SessionRecord],
-        sessions: [SessionData],
-        claudeMetadata: ClaudeDesktopSessionMetadataSnapshot?,
-        codexThreads: any CodexThreadStateProviding,
-        now: Date
-    ) -> SessionClassificationSnapshot {
-        let externallyClassifiableSessions = sessions.filter { !$0.hidden && !$0.shouldAutoHide }
-        let codexDelegationIndex = codexThreads.stateIndex(
-            matching: codexThreadIDs(in: sessions)
-        )
-        let archivedCodexThreadIDs = codexDelegationIndex?.archivedThreadIDs.intersection(
-            codexThreadIDs(in: sessions)
-        ) ?? []
-        let missingCodexThreadIDs = missingCodexThreadIDs(
-            in: externallyClassifiableSessions,
-            index: codexDelegationIndex,
-            now: now
-        )
-        let codexInternalHelperThreadIDs = codexDelegationIndex?.internalHelperThreadIDs ?? []
-        let uncertainCodexDelegationThreadIDs = codexDelegationIndex?.uncertainDelegationThreadIDs ?? []
-        let contradictoryCodexDelegationThreadIDs = codexDelegationIndex?.contradictoryDelegationThreadIDs ?? []
-        let codexExecHelperThreadIDs = codexExecHelperThreadIDs(
-            in: externallyClassifiableSessions,
-            index: codexDelegationIndex
-        )
-        let archivedClaudeSessionIDs = claudeMetadata?.archivedSessionIDs ?? []
-
-        let evidence = SessionClassificationEvidence(
-            archivedCodexThreadIDs: archivedCodexThreadIDs,
-            missingCodexThreadIDs: missingCodexThreadIDs,
-            codexInternalHelperThreadIDs: codexInternalHelperThreadIDs,
-            uncertainCodexDelegationThreadIDs: uncertainCodexDelegationThreadIDs,
-            contradictoryCodexDelegationThreadIDs: contradictoryCodexDelegationThreadIDs,
-            codexExecHelperThreadIDs: codexExecHelperThreadIDs,
-            archivedClaudeSessionIDs: archivedClaudeSessionIDs
-        )
-        let records = candidates.map { candidate in
-            ClassifiedSessionRecord(
-                url: URL(fileURLWithPath: candidate.path),
-                candidate: candidate,
-                disposition: disposition(
-                    for: candidate.data,
-                    evidence: evidence,
-                    claudeMetadata: claudeMetadata
-                )
-            )
-        }
-        return SessionClassificationSnapshot(records: records, evidence: evidence)
-    }
-
-    private nonisolated static func disposition(
-        for data: SessionData,
-        evidence: SessionClassificationEvidence,
-        claudeMetadata: ClaudeDesktopSessionMetadataSnapshot?
-    ) -> SessionDisposition {
-        // Priority is behavior-bearing: Codex archive wins so no other hidden reason can protect
-        // an archived path from Cleanup. Durable local hides then precede external missing/helper state.
-        let isArchivedCodexSession = data.isCodex
-            && evidence.archivedCodexThreadIDs.contains(data.sessionId)
-        if isArchivedCodexSession {
-            return .hidden(.archivedCodexThread)
-        }
-        if data.hidden {
-            return .hidden(.persistedHidden)
-        }
-        if data.shouldAutoHide {
-            return .hidden(.autoHidden)
-        }
-        if isMissingCodexSession(data, missingThreadIDs: evidence.missingCodexThreadIDs) {
-            return .hidden(.missingCodexThread)
-        }
-        if isCodexInternalHelperSession(data, internalHelperThreadIDs: evidence.codexInternalHelperThreadIDs) {
-            return .hidden(.codexInternalHelper)
-        }
-        if isCodexExecHelperSession(data, execHelperThreadIDs: evidence.codexExecHelperThreadIDs) {
-            return .hidden(.codexExecHelper)
-        }
-        if isArchivedClaudeDesktopSession(data, archivedSessionIDs: evidence.archivedClaudeSessionIDs) {
-            return .hidden(.archivedClaudeDesktop)
-        }
-        // Claude Desktop `SessionEnd` is worker/session termination, not archive; matched metadata stays resumable.
-        if isOrphanedEndedClaudeDesktopSession(data, metadataSnapshot: claudeMetadata) {
-            return .hidden(.orphanedEndedClaudeDesktop)
-        }
-        if isClaudeDesktopStartupPlaceholder(data, metadataSnapshot: claudeMetadata) {
-            return .hidden(.claudeDesktopStartupPlaceholder)
-        }
-        return .display
-    }
-
-    func deriveSessionClassification(from decoded: [(url: URL, session: SessionData)]) -> SessionClassificationSnapshot {
-        let now = dataSources.now()
-        let codexThreads = batchedCodexThreadState(
-            in: decoded.map(\.session),
-            codexThreads: dataSources.codexThreads
-        )
-        let repaired = repairStickyCodexDelegationState(in: decoded, codexThreads: codexThreads)
-        let classifiableSessions = repaired
-            .map(\.session)
-            .filter { !$0.hidden && !$0.shouldAutoHide }
-        let claudeMetadata = Self.claudeDesktopMetadataSnapshot(
-            in: classifiableSessions,
-            claudeDesktopSessions: dataSources.claudeDesktopSessions
-        )
-        let candidates = Self.buildCandidates(
-            repaired,
-            now: now,
-            desktopAppConnectionLookup: dataSources.desktopAppConnection,
-            claudeMetadata: claudeMetadata,
-            processAlive: dataSources.processAlive
-        )
-        return Self.sessionClassificationSnapshot(
-            in: candidates,
-            sessions: repaired.map(\.session),
-            claudeMetadata: claudeMetadata,
-            codexThreads: codexThreads,
-            now: now
-        )
-    }
-
-    private func repairStickyCodexDelegationState(
-        in decoded: [(url: URL, session: SessionData)],
-        codexThreads: FrozenCodexThreadState
-    ) -> [(url: URL, session: SessionData)] {
-        guard let index = codexThreads.index else { return decoded }
-        let repairableThreadIDs = index.interactiveRootThreadIDs
-            .intersection(index.knownNoSpawnEdgeThreadIDs)
-        guard !repairableThreadIDs.isEmpty else { return decoded }
-
-        var repairedByPath: [String: SessionData] = [:]
-        for (url, data) in decoded where data.hidden
-            && data.isSubagentSession
-            && repairableThreadIDs.contains(data.sessionId) {
-            withSessionLockForMaintenance(
-                sessionPath: url.path,
-                sessionId: data.sessionId,
-                action: "Codex delegated-task classification repair"
-            ) {
-                guard let repaired = try Self.repairedCodexInteractiveRootSessionSnapshot(
-                    path: url.path,
-                    repairableThreadIDs: repairableThreadIDs
-                ) else {
-                    return
-                }
-                try repaired.writeToFile(path: url.path)
-                repairedByPath[url.path] = repaired
-                sessionManagerLogger.info(
-                    "restoring interactive Codex session \(data.sessionId, privacy: .public)"
-                )
-            }
-        }
-
-        return decoded.map { entry in
-            (entry.url, repairedByPath[entry.url.path] ?? entry.session)
-        }
-    }
-
-    private func batchedCodexThreadState(
-        in sessions: [SessionData],
-        codexThreads: any CodexThreadStateProviding
-    ) -> FrozenCodexThreadState {
-        let threadIDs = Self.codexThreadIDs(in: sessions)
-        let snapshot = codexThreads.stateSnapshot(matching: threadIDs)
-        let effectiveIndex = codexThreadClassificationMemory.effectiveIndex(
-            from: snapshot,
-            matching: threadIDs
-        )
-        return FrozenCodexThreadState(index: effectiveIndex)
-    }
-
-    private nonisolated static func codexThreadIDs(in sessions: [SessionData]) -> Set<String> {
-        Set(
-            sessions
-                .filter(\.isCodex)
-                .map(\.sessionId)
-        )
-    }
-
-    /// Codex sessions should correspond to a Codex thread row. If the thread store is
-    /// readable and the row is gone, the app should not publish the stale hook session.
-    private nonisolated static func missingCodexThreadIDs(
-        in sessions: [SessionData],
-        index: CodexThreadStateIndex?,
-        now: Date
-    ) -> Set<String> {
-        guard let index else { return [] }
-        let codexSessions = sessions.filter(\.isCodex)
-        let threadIDs = Set(codexSessions.map(\.sessionId))
-        let resolvedThreadIDs = threadIDs.subtracting(index.unknownThreadIDs)
-        let missingThreadIDs = resolvedThreadIDs.subtracting(index.existingThreadIDs)
-        let freshMissingThreadIDs = Set(codexSessions.compactMap { session -> String? in
-            guard missingThreadIDs.contains(session.sessionId),
-                  now.timeIntervalSince(session.lastActivity) <= Self.codexMissingThreadGraceSeconds else {
-                return nil
-            }
-            return session.sessionId
-        })
-        return missingThreadIDs.subtracting(freshMissingThreadIDs)
-    }
-
-    nonisolated static func isMissingCodexSession(
-        _ session: SessionData,
-        missingThreadIDs: Set<String>
-    ) -> Bool {
-        session.isCodex && missingThreadIDs.contains(session.sessionId)
-    }
-
-    nonisolated static func isCodexInternalHelperSession(
-        _ session: SessionData,
-        internalHelperThreadIDs: Set<String>
-    ) -> Bool {
-        session.isCodex && internalHelperThreadIDs.contains(session.sessionId)
-    }
-
-    /// Codex can launch short-lived `codex exec` helper threads. They are useful as
-    /// rollout artifacts but should not appear as user-visible cctop sessions.
-    private nonisolated static func codexExecHelperThreadIDs(
-        in sessions: [SessionData],
-        index: CodexThreadStateIndex?
-    ) -> Set<String> {
-        let threadIDs = codexThreadIDs(in: sessions)
-        return index?.execHelperThreadIDs.intersection(threadIDs) ?? []
-    }
-
-    nonisolated static func isCodexExecHelperSession(
-        _ session: SessionData,
-        execHelperThreadIDs: Set<String>
-    ) -> Bool {
-        session.isCodex && execHelperThreadIDs.contains(session.sessionId)
-    }
-
-    /// Re-read the session file under its lock before persisting a hidden flag. The helper IDs
-    /// come from the same batched authoritative snapshot used for classification, avoiding a
-    /// second database query per session while still rejecting a concurrently replaced file.
-    nonisolated static func codexInternalHelperHiddenSessionSnapshot(
-        path: String,
-        internalHelperThreadIDs: Set<String>
-    ) throws -> SessionData? {
-        guard FileManager.default.fileExists(atPath: path) else { return nil }
-        var latest = try SessionData.fromFile(path: path)
-        guard !latest.hidden, latest.isCodex else { return nil }
-        guard internalHelperThreadIDs.contains(latest.sessionId) else { return nil }
-        latest.isSubagentSession = true
-        latest.hidden = true
-        return latest
-    }
-
-    /// Repairs only the stale marker pair written by the old `thread_source` classifier. The
-    /// caller supplies a batched, authoritative set that has already proven `cli`/`vscode` and
-    /// a readable topology table with no spawn edge.
-    nonisolated static func repairedCodexInteractiveRootSessionSnapshot(
-        path: String,
-        repairableThreadIDs: Set<String>
-    ) throws -> SessionData? {
-        guard FileManager.default.fileExists(atPath: path) else { return nil }
-        var latest = try SessionData.fromFile(path: path)
-        guard latest.hidden,
-              latest.isSubagentSession,
-              latest.isCodex,
-              repairableThreadIDs.contains(latest.sessionId) else {
-            return nil
-        }
-
-        latest.isSubagentSession = false
-        guard !latest.shouldAutoHide else { return nil }
-        latest.hidden = false
-        return latest
-    }
-
-    nonisolated static func autoHiddenSessionSnapshot(path: String) throws -> SessionData? {
-        guard FileManager.default.fileExists(atPath: path) else { return nil }
-        var latest = try SessionData.fromFile(path: path)
-        guard !latest.hidden, latest.shouldAutoHide else { return nil }
-        latest.hidden = true
-        return latest
-    }
-
-    func hideCodexInternalHelperSessions(_ candidates: [SessionRecord]) {
-        let internalHelperThreadIDs = Set(candidates.map(\.data.sessionId))
-        for candidate in candidates {
-            withSessionLockForMaintenance(
-                sessionPath: candidate.path,
-                sessionId: candidate.data.sessionId,
-                action: "Codex internal helper hide"
-            ) {
-                guard let hiddenSession = try Self.codexInternalHelperHiddenSessionSnapshot(
-                    path: candidate.path,
-                    internalHelperThreadIDs: internalHelperThreadIDs
-                ) else {
-                    return
-                }
-                sessionManagerLogger.info(
-                    "hiding Codex internal helper session \(candidate.data.sessionId, privacy: .public)"
-                )
-                try hiddenSession.writeToFile(path: candidate.path)
-            }
-        }
-    }
-
-    nonisolated static func claudeDesktopMetadataSnapshot(
-        in sessions: [SessionData],
-        claudeDesktopSessions: any ClaudeDesktopSessionStateProviding = ClaudeDesktopSessionArchiveLookup()
-    ) -> ClaudeDesktopSessionMetadataSnapshot? {
-        let sessionIDs = Set(
-            sessions
-                .filter(\.isClaudeDesktopHost)
-                .map(\.sessionId)
-        )
-        return claudeDesktopSessions.metadataSnapshot(matching: sessionIDs)
-    }
-
-    nonisolated static func isArchivedClaudeDesktopSession(
-        _ session: SessionData,
-        archivedSessionIDs: Set<String>
-    ) -> Bool {
-        session.isClaudeDesktopHost && archivedSessionIDs.contains(session.sessionId)
-    }
-
-    nonisolated static func isOrphanedEndedClaudeDesktopSession(
-        _ session: SessionData,
-        metadataSnapshot: ClaudeDesktopSessionMetadataSnapshot?
-    ) -> Bool {
-        guard session.isClaudeDesktopHost,
-              session.endedAt != nil || session.disconnectedAt != nil,
-              metadataSnapshot?.isAuthoritative == true else {
-            return false
-        }
-        return metadataSnapshot?.matchedSessionIDs.contains(session.sessionId) == false
-    }
-
-    nonisolated static func isClaudeDesktopStartupPlaceholder(
-        _ session: SessionData,
-        metadataSnapshot: ClaudeDesktopSessionMetadataSnapshot?
-    ) -> Bool {
-        guard session.isClaudeDesktopHost,
-              session.endedAt == nil, session.disconnectedAt == nil,
-              metadataSnapshot?.isAuthoritative == true, metadataSnapshot?.matchedSessionIDs.contains(session.sessionId) == false,
-              session.status == .idle,
-              isBlank(session.sessionName),
-              isBlank(session.lastPrompt),
-              isBlank(session.lastTool),
-              isBlank(session.lastToolDetail),
-              isBlank(session.notificationMessage),
-              session.activeSubagents?.isEmpty ?? true else {
-            return false
-        }
-        return true
-    }
-    private nonisolated static func isBlank(_ value: String?) -> Bool {
-        value?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty ?? true
     }
 
     /// Fresh single-session archive check for the GC deletion decision. Unlike the batch snapshot
